backend/requirements.txt
anthropic
fastapi
google-generativeai
google-adk
litellm
numpy
pandas
pydantic
pydantic-settings
python-dotenv
requests
uvicorn
yfinance
zhipuai
scikit-learn
pytest
pytest-asyncio
aiohttp
pdfplumber>=0.10.0
python-multipart>=0.0.6
openpyxl>=3.1.0
pytesseract>=0.3.10
pillow>=10.0.0
opencv-python>=4.8.0
